+++ src/Utils/MiscUI/MyGraph.cpp
@@ -177,7 +177,7 @@ int MyGraphSeries::GetAverageDataValue() const
 	if (m_dwaValues.GetSize() == 0)
 		return 0;
 
-	return nTotal / m_dwaValues.GetSize();
+	return nTotal / (int)m_dwaValues.GetSize();
 }
 
 // Returns the number of data points that are not zero.
@@ -1214,8 +1214,8 @@ void MyGraph::DrawSeriesBar(CDC& dc) const
 		int nMaxDataValue = max(GetMaxDataValue(), 1);
 		double barTop = m_ptOrigin.y - (double)m_nYAxisHeight *
 			(GetAverageDataValue() / (double)nMaxDataValue);
-		dc.MoveTo(m_ptOrigin.x, barTop);
-		VERIFY(dc.LineTo(m_ptOrigin.x + (m_nXAxisWidth - m_rcLegend.Width() - (GAP_PIXELS * 2)), barTop));
+		dc.MoveTo(m_ptOrigin.x, (int)barTop);
+		VERIFY(dc.LineTo(m_ptOrigin.x + (m_nXAxisWidth - m_rcLegend.Width() - (GAP_PIXELS * 2)), (int)barTop));
 	}
 }
 
@@ -1317,8 +1317,8 @@ void MyGraph::DrawSeriesLine(CDC& dc) const
 	int nMaxDataValue = max(GetMaxDataValue(), 1);
 	double barTop = m_ptOrigin.y - (double)m_nYAxisHeight *
 		(GetAverageDataValue() / (double)nMaxDataValue);
-	dc.MoveTo(m_ptOrigin.x, barTop);
-	VERIFY(dc.LineTo(m_ptOrigin.x + (m_nXAxisWidth - m_rcLegend.Width() - (GAP_PIXELS * 2)), barTop));
+	dc.MoveTo(m_ptOrigin.x, (int)barTop);
+	VERIFY(dc.LineTo(m_ptOrigin.x + (m_nXAxisWidth - m_rcLegend.Width() - (GAP_PIXELS * 2)), (int)barTop));
 }
 
 //
